# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "AreaResponse": {
        "properties": {
          "area_type": {
            "const": "commune",
            "title": "Area Type",
            "type": "string"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "type": "array"
          },
          "center": {
            "items": {
              "type": "number"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Center",
            "type": "array"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "department_code": {
            "title": "Department Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "area_type",
          "code",
          "name",
          "department_code",
          "center",
          "bbox"
        ],
        "title": "AreaResponse",
        "type": "object"
      },
...
      "CoverageResponse": {
        "properties": {
          "acceptance_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Status"
          },
          "coverage_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Ratio"
          },
          "data_source_id": {
            "title": "Data Source Id",
            "type": "string"
          },
          "freshest_observation_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshest Observation At"
          },
          "matched_record_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Record Count"
          },
          "measured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured At"
          },
          "record_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Count"
          },
          "release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Id"
          },
          "source_published_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Published On"
          }
        },
        "required": [
          "data_source_id",
          "release_id",
          "acceptance_status",
          "source_published_on",
          "record_count",
          "matched_record_count",
          "coverage_ratio",
          "freshest_observation_at",
          "measured_at"
        ],
        "title": "CoverageResponse",
        "type": "object"
      },
...
      "EntityDetailResponse": {
        "properties": {
          "area_m2": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area M2"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "type": "array"
          },
          "center": {
            "items": {
              "type": "number"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Center",
            "type": "array"
          },
          "commune_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune Code"
          },
          "commune_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune Name"
          },
          "department_code": {
            "title": "Department Code",
            "type": "string"
          },
          "entity_type": {
            "enum": [
              "parcel",
              "building",
              "property_unit"
            ],
            "title": "Entity Type",
            "type": "string"
          },
          "geometry": {
            "additionalProperties": true,
            "title": "Geometry",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "related_entities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Related Entities",
            "type": "array"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "id",
          "entity_type",
          "label",
          "commune_code",
          "commune_name",
          "department_code",
          "center",
          "bbox",
          "geometry",
          "properties",
          "related_entities",
          "sources"
        ],
        "title": "EntityDetailResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MarketContextResponse": {
        "properties": {
          "comparables": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Comparables",
            "type": "array"
          },
          "energy_assessment": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Assessment"
          },
          "property_unit_id": {
            "title": "Property Unit Id",
            "type": "string"
          },
          "risks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Risks",
            "type": "array"
          },
          "snapshot_at": {
            "title": "Snapshot At",
            "type": "string"
          },
          "urban_zone": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urban Zone"
          }
        },
        "required": [
          "property_unit_id",
          "snapshot_at",
          "comparables",
          "energy_assessment",
          "urban_zone",
          "risks"
        ],
        "title": "MarketContextResponse",
        "type": "object"
      },
...
      "PropertyUnitSummaryResponse": {
        "properties": {
          "area_m2": {
            "minimum": 0.0,
            "title": "Area M2",
            "type": "number"
          },
          "building_count": {
            "minimum": 0.0,
            "title": "Building Count",
            "type": "integer"
          },
          "building_footprint_m2": {
            "minimum": 0.0,
            "title": "Building Footprint M2",
            "type": "number"
          },
          "cadastral_id": {
            "title": "Cadastral Id",
            "type": "string"
          },
          "center": {
            "items": {
              "type": "number"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Center",
            "type": "array"
          },
          "commune_code": {
            "title": "Commune Code",
            "type": "string"
          },
          "commune_name": {
            "title": "Commune Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cadastral_id",
          "commune_code",
          "commune_name",
          "area_m2",
          "building_count",
          "building_footprint_m2",
          "center"
        ],
        "title": "PropertyUnitSummaryResponse",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "bbox": {
            "items": {
              "type": "number"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "type": "array"
          },
          "center": {
            "items": {
              "type": "number"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Center",
            "type": "array"
          },
          "entity_type": {
            "enum": [
              "address",
              "area",
              "parcel"
            ],
            "title": "Entity Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "secondary_label": {
            "title": "Secondary Label",
            "type": "string"
          }
        },
        "required": [
          "entity_type",
          "id",
          "label",
          "secondary_label",
          "center",
          "bbox"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ViewportResponse": {
        "properties": {
          "coverage": {
            "enum": [
              "covered",
              "outside_coverage"
            ],
            "title": "Coverage",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PropertyUnitSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "partial": {
            "title": "Partial",
            "type": "boolean"
          }
        },
        "required": [
          "coverage",
          "partial",
          "items"
        ],
        "title": "ViewportResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Immo Opportunities API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/areas": {
      "get": {
        "operationId": "areas_api_v1_areas_get",
        "parameters": [
          {
            "in": "query",
            "name": "department_code",
            "required": false,
            "schema": {
              "default": "35",
              "pattern": "^[0-9A-Z]{2,3}$",
              "title": "Department Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "area_type",
            "required": false,
            "schema": {
              "const": "commune",
              "default": "commune",
              "title": "Area Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AreaResponse"
                  },
                  "title": "Response Areas Api V1 Areas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Areas",
        "tags": [
          "real-map"
        ]
      }
    },
    "/api/v1/buildings/{building_id}": {
      "get": {
        "operationId": "building_api_v1_buildings__building_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "building_id",
            "required": true,
            "schema": {
              "title": "Building Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Building",
        "tags": [
          "real-map"
        ]
      }
    },
...
    "/api/v1/market-data/coverage": {
      "get": {
        "operationId": "market_data_coverage_api_v1_market_data_coverage_get",
        "parameters": [
          {
            "in": "query",
            "name": "commune_code",
            "required": true,
            "schema": {
              "pattern": "^[0-9A-Z]{5}$",
              "title": "Commune Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CoverageResponse"
                  },
                  "title": "Response Market Data Coverage Api V1 Market Data Coverage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Data Coverage",
        "tags": [
          "market-data"
        ]
      }
    },
...
    "/api/v1/parcels/{parcel_id}": {
      "get": {
        "operationId": "parcel_api_v1_parcels__parcel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "parcel_id",
            "required": true,
            "schema": {
              "title": "Parcel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parcel",
        "tags": [
          "real-map"
        ]
      }
    },
    "/api/v1/property-units": {
      "get": {
        "operationId": "property_units_api_v1_property_units_get",
        "parameters": [
          {
            "in": "query",
            "name": "west",
            "required": true,
            "schema": {
              "title": "West",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "south",
            "required": true,
            "schema": {
              "title": "South",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "east",
            "required": true,
            "schema": {
              "title": "East",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "north",
            "required": true,
            "schema": {
              "title": "North",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 250,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Property Units",
        "tags": [
          "real-map"
        ]
      }
    },
    "/api/v1/property-units/{property_unit_id}": {
      "get": {
        "operationId": "property_unit_api_v1_property_units__property_unit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "property_unit_id",
            "required": true,
            "schema": {
              "title": "Property Unit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Property Unit",
        "tags": [
          "real-map"
        ]
      }
    },
    "/api/v1/property-units/{property_unit_id}/market-context": {
      "get": {
        "operationId": "property_unit_market_context_api_v1_property_units__property_unit_id__market_context_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_unit_id",
            "required": true,
            "schema": {
              "title": "Property Unit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "snapshot_at",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Snapshot At",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Property Unit Market Context",
        "tags": [
          "market-data"
        ]
      }
    },
...
    "/api/v1/search": {
      "get": {
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 3,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "department_code",
            "required": false,
            "schema": {
              "default": "35",
              "pattern": "^[0-9A-Z]{2,3}$",
              "title": "Department Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 30,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResponse"
                  },
                  "title": "Response Search Api V1 Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "real-map"
        ]
      }
    },
...
  }
}